Some refactor and some maths

+++ src/main/java/losango/domain/Coordinate.java
@@ -5,35 +5,31 @@ package losango.domain;
  */
 public class Coordinate {
 
-    private int latitude;
-    private int longitude;
+    private double latitude;
+    private double longitude;
 
     public Coordinate() {
     }
 
-    public Coordinate(int latitude, int longitude) {
+    public Coordinate(double latitude, double longitude) {
         this.latitude = latitude;
         this.longitude = longitude;
     }
 
-    public int getLatitude() {
+    public double getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(int latitude) {
+    public void setLatitude(double latitude) {
         this.latitude = latitude;
     }
 
-    public int getLongitude() {
+    public double getLongitude() {
         return longitude;
     }
 
-    public void setLongitude(int longitude) {
+    public void setLongitude(double longitude) {
         this.longitude = longitude;
     }
-
-
-
-
 
 }
